automate-chef-io: configuration+LDAP: use warning shortcodes using {{% .. %}}
Otherwise, the text in the shortcode would be taken as-is. The text we
sometimes put in there, however, contains links and backticks, and
SHOULD be processed.

Since I believe the verbatim variant should be the exception, I've
flipped them all.

I've checked the other occurrances of `{{< warning >}}` in our docs and
they don't seem to have any backticks or links etc. So I haven't changed
them.

Signed-off-by: Stephan Renatus <srenatus@chef.io>