Fix race condition in the grpc cereal backend (#1204)
* Fix race condition in the grpc cereal backend

Task completion wasn't waiting around for the server to confirm that it
processed the task completion messages. This was causing the tests to
fail as they assumed that once the backend function returns (without
error), the data has been persisted.

Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>

* Try a different type of fix

The last fix created 2 readers on the stream which was causing things to
hang. Instead, the cancelation thread will decide when we're done. It
will pass the error through a channel

Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>