Preventing empty error message suggestions (#1235)
Preventing empty error message suggestions. The problem was that the errorMessage field was filled with an empty string if there was no error. Therefore the empty string suggestion was correct. The bug was prevented in three places ingestion, request to the API from the UI, and the search bar component.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>