[lib/cereal] Avoid spurious connection closes
Calling cancel() on an uncommitted/rolledback transactions results in
the underlying connection also being closed.

This happens in the common case of the workflow scheduler when there
are no scheduled workflows to run. To avoid log messages for
automate-pg-gateway about the connection close, we explicitly commit
the transaction to avoid it.

Signed-off-by: Steven Danna <steve@chef.io>