[bldr-config-gen] Individually track vendored deps
Previously, any component that used a dependency simply tracked any
changes to Gopkg.lock. As of this change, we specifically add globs
for the things you depend on.

The good thing about this change is that if you update Gopkg.lock you
don't necessarily have to rebuild the entire world.

The bad thing about this change is that you'll need to regenerate
.bldr.toml anytime you add a dependency to your project.

Signed-off-by: Steven Danna <steve@chef.io>