Fixing event feed filtering (#1218)
Changing event feed filters to filter on fields and not tags. The event feed service was performing filters with 'event type', 'requestor', and 'organizations' on the tags field. This does not work because if any of the above filters have similar values, the filtered events would have been incorrectly returned. For example, if an event had a 'requestor' = 'profile' and 'event type' = 'scan job' then when filtering for 'event type' = 'profile' this event would be returned incorrectly.

The Change required a migration of the eventfeed-1-feeds Elasticsearch index. Because of the migration, other mapping changes were made to prevent future migrations.

Pull out the FormatFilters function from three components (applications, event-feed, and config-mgmt) into the lib directory.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>