Dan/convert to denormalized schema (#1257)
Change the schema of the applications service to a single table for all service/service group/supervisor/deployment data. This makes the data model more closely match the domain operations in the application. Currently data is only created via health check message with full metadata, and entities such as deployments, service groups, and supervisors exist by having at least one service that includes them in the metadata.

Specific benefits of this change are:

*health check messages can be ingested with one Upsert query, which is much simpler than what we do now.
* delete (for disconnected services) can just delete the relevant service rows.
* query to power the applications UI will become simpler. Currently it uses a variable number of joins based on filter criteria, we can get rid of all the joins.

Some of the typical downsides of denormalizing are not present:

we don't expose any way to update any of the attributes that apply to aggregations of services (e.g., rename an application or environment) as a group; instead users would reconfigure all of the supervisors running those services. So we don't have any fan-out edits we need to worry about.

It's almost certain that this change will increase disk usage a little bit, but disk usage isn't the limiting factor of our scalability, so the reduction in complexity is worth it.

Overall, we expect this change to make it easier to develop new features in the future due to a reduction in code and code complexity.