deployment-service: fix erroneous append
Uses of append that aren't in the form of

    someSlice = append(someSlice, items...)

or

    someSlice := append([]T{someT1, sometTN}, items...)

should probably be looked at with suspicion because append modifies
the slice's backing array and it can easily produce unexpected
results.

This seems to sneak in during cleanups or when trying to make code a
bit less verbose.

In the case here, we can just avoid the append completely since we
were just doing it to more easily iterate over a set of services.
Since the only addition to the ExpectedServices array was the
deployment-service, I've opted to just add the deployment-service
explicitly after the loop.

Signed-off-by: Steven Danna <steve@chef.io>