Fix restoring old backups with latest A2 (#1262)
The current version of automate was not able to restore backups from some older version. This is because we updated the pg client tools to 11 to support HA, but those tools don't quite support backups taken with pg 10 tools. We strip out the problematic `CREATE SCHEMA public` stuff which now happens outside of backups. This prevents `pg_restore` from complaining about that schema already existing.
Also added a test for this case