lib/cereal: Use in-process notifications to avoid delays (#1255)
This change allows us to wake up relevant go-routines in the local
process when certain events happens:

When we enqueue tasks, we wake up the task polling goroutine
When we finish a task, we wake up the workflow polling goroutine
For each registered task executor we have a "Poller" which wakes up at an interval
or on notification and starts a task worker. An additional worker will be spawned as
long as there is more work to do. When the worker finds that there is no more work to
do it exits.

Signed-off-by: Steven Danna <steve@chef.io>