Fix race condition in disabled status of team details projects dropdown
The dropdown was not becoming properly disabled when saves were in progress, resulting in a race condition that made the tests flaky. Also hardened the cypress tests by adding 'is.not.disabled' assertion so that cypress will actually wait for the dropdown to not be disabled before sending the click event.

Signed-off-by: Tyler Cloke <tylercloke@gmail.com>