[compliance] Use the "cereal" library for scan jobs (#595)
To facilitate multi-node operations, this moves us to using the
cereal library to run compliance scan jobs. With this change,
scan job tasks will

- Run across multiple compliance-service instances talking to the
   same database, and Safe to resume after restart

- To do this, the scan job process, has to be re-formulated into
   a set of tasks. In doing this, we've attempted to change as
   little of the core scan jobs logics as possible.

Currently, we are not yet using the built-in scheduler in the
cereal library. We've done this to avoid a data migration for
now. Rather we rely on the fact that the library prevents two
instances of the same workflow from running at once. Thus, even
if two nodes try to independently schedule a job only one should
win. Since the scheduled_time of the workflow is modified as part
of the workflow, we think that this is also safe against cases
that would otherwise result in duplicate scans. There is still a
window where we could end up with a duplicate job:

- Thread A gets Job 1
- Thread B gets Job 1
- Thread A pushes and finishes Job 1
- Thread B pushes and finished Job 1

But the existing code already has a number of bugs where a job
can get run twice so we are defering working on this for the
moment.

Signed-off-by: Steven Danna <steve@chef.io>
Signed-off-by: Alex Pop <alexpop@users.noreply.github.com>
Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>
Co-authored-by: Alex Pop <alexpop@users.noreply.github.com>
Co-authored-by: Jay Mundrawala <jmundrawala@chef.io>