Fix race condition in cypress test by preventing typing in input before team loaded
If you have a team in the store already, it will populate the team save input before the team is fetched via the API. Since we didn't prevent typing in the input before the team was fetched via the API, cypress would find the input, find that the team name was correct (from the store cache), and then type. However, this would happen so fast the team wasn't fetched yet. When the fetch completed, it overwrote the updated team name cypress typed before save was clicked.

I've added functionality to disable the team name input before the team is fetched and while a save is in progress. The cypress tests should now pass since they are asserting that the input is enabled, preventing the race condition.

Signed-off-by: Tyler Cloke <tylercloke@gmail.com>