Add cereal committed messages (#1214)
Previously we were relying on closing streams for
completion. This was turning out to be a little bit too
magical. The clients now wait for a message that says what they
have done has been committed (or not). For tasks, we also no
longer log an error every time we successfully complete a task
because the stream got closed and we're unsure if we got canceled
or we finished.

Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>