Wire up override config for restore (#113)
Add the ability to pass a patch config to the restore process with `--patch-config`. The provided config will be layered on top of the config that is in the backup. This is useful for migrating from the internal data services to external and for restoring into a new data services cluster

We can also provide the exact config `--set-config`.