lib/cereal: Set MaxIdleConnections in postgresql backend to 4
In an ideal world, we would not be managing this database object
ourselves. However, since we are, this sets the MaxIdleConnections to
a value that we think is a bit more appropriate for our current users
of cereal.

The end goal of this is to cut down on the log spam caused by
perpetually closing connections.

Signed-off-by: Steven Danna <steve@chef.io>