lower poll interval for tests
These tests broke when jittering the timer because the larger polling
interval we are using now meant that if we woke up before the delayed
task on the first wakeup, we would miss the required interval on the
second wakeup. Lowering the interval more in line with the test
assertions fixes the test failure.

Signed-off-by: Steven Danna <steve@chef.io>