Use context library from standard lib
This is just a small cleanup I ran. All of the import changes should
be a no-op because the x/net/context package simply points to the
standard library when building on a recent version of Go, but keeping
these imports consistent with everything else seems nice.

For trial-license-service, I also removed use of the ctxhttp package
in favor of the context support in the standard library.

Signed-off-by: Steven Danna <steve@chef.io>