[lib/cereal] Add Chain Workflow pattern
The chain workflow pattern is usued to be able to combine two workflow
executors to run one after the other. This runs both in the same
workflow instance. Using this pattern is a good way to take a large
workflow and break in down into smaller workflows.

You may not nest chains as the level of magic needed to make that work is
quite high and this PR is already getting big.

This PR also adds a unit testing framework for cereal workflows. It's
not perfect. There is a bit of weirdness around Decisions. We currently
test the last decision the workflow instance returned instead of the
decision the workflow executor's callback returned. We need to do this
until Decision can be more easily unit tested. It currently holds a lot
of private bits (the tasks that will be enqueued).

Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>