Add workflow for project rules update
This PR replaces the current project update system with one driven by
workflows in cereal. The workflow is driven by the authz-service and
consists of 2 steps: First, authz commits the staged rules. After that
is done, the domain services update is started and pinged for status
every 10 seconds. These are written as 2 different workflows and tied
together using a combination of the Chain workflow pattern and Parallel
workflow pattern.

Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>