automate-ui: flip defaults for chef-session's logout()
Before, we'd have to pass in a url and a boolean for having the proper
page to go back to, and controling sign in method selection.

In the request interceptor, we don't know which page to go back to --
but chef-session already has a kludgy way of knowing. So, the url
parameter is now optional, and if it is not provided, will default to
that.

The only place where we want to override both is the explicit logout
from the profile dropdown.

Signed-off-by: Stephan Renatus <srenatus@chef.io>