Event Feed Filtering not working for Profiles and Scan Jobs (#1358)
Filtering scan job events was not working on the event feed page because the filter tag for scan jobs was incorrect for the event-feed-service API. We were requesting 'scan_job' but it should be 'scanjobs'
The right grouped sidebar was not displaying profiles or scan jobs for two reasons. The first reason was that the event type tag was incorrect. Second, the event feed was looking for the user tag to be 'requestorName' by it should be 'requestor_name'.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>