tools/cereal-scaffold: separate enqueue and starting
This separates the enqueue operation into a new command. This allows
you to more easily start multiple cereal managers and then enqueue a
single task. One downside of this is that since these operations live
in different processes, we don't benefit from the in-process
signaling.

I've also changed a bit about how the simple-workflow total time
tracking works.

Alternatively, maybe it is time to nuke this command? I still find it
useful for manually testing changes, but we now have a more complete
test suite and we should probably use that instead.

Signed-off-by: Steven Danna <steve@chef.io>