lib/cereal limit number of concurrent TaskDequeue calls (#1334)
For each registered task, we call TaskDequeue every
TaskPoolInterval. For a service with a large number of task executors,
this can cause a large number TaskDequeue calls to be made to the
backend.

Similarly, even with 1 TaskExecutor configured for 10 workers, if
those workers are processing a large number of tasks, each worker will
call TaskDequeue. This may not be the most efficient way to use a
given backend.

This changes adds a small pool that can be used to limit the
concurrency of TaskDequeue calls.

Signed-off-by: Steven Danna <steve@chef.io>