Project Update Workflow and patterns to support it (#1346)
* [lib/cereal] Add Chain Workflow pattern

The chain workflow pattern is usued to be able to combine two workflow
executors to run one after the other. This runs both in the same
workflow instance. Using this pattern is a good way to take a large
workflow and break in down into smaller workflows.

You may not nest chains as the level of magic needed to make that work is
quite high and this PR is already getting big.

This PR also adds a unit testing framework for cereal workflows. It's
not perfect. There is a bit of weirdness around Decisions. We currently
test the last decision the workflow instance returned instead of the
decision the workflow executor's callback returned. We need to do this
until Decision can be more easily unit tested. It currently holds a lot
of private bits (the tasks that will be enqueued).

* [lib/cereal] Add Parallel Workflow pattern

The parallel workflow pattern can be used to combine multiple workflow
executors to run in parallel. These subworkflows run under one workflow
instance.

You may not nest parallel workflow executors.

See #1347

* Add workflow for project rules update

This PR replaces the current project update system with one driven by
workflows in cereal. The workflow is driven by the authz-service and
consists of 2 steps: First, authz commits the staged rules. After that
is done, the domain services update is started and pinged for status
every 10 seconds. These are written as 2 different workflows and tied
together using a combination of the Chain workflow pattern and Parallel
workflow pattern.

See #1349