ci: purge fastly path during manifest creation
The Habitat API has at least 2 layers of caching for the /latest
endpoint in use in create-manifest.rb. 1 layer of caching is in the
service itself. This cache should be invalidated on write.

The second layer of cache is via Fastly.

We have created incorrect manifests on multiple occasions and we
currently believe it is the result of this caching.

In the long-run, we would like to use the expeditor workload directly;
however, a bug in expeditor currently prevents that.

This PR purges fastly for expeditor-based builds in the hopes that it
is the fastly cache that is the cause of our inconsistent manifests.

Signed-off-by: Steven Danna <steve@chef.io>