scaffolding: Always add sys.ip to NOPROXY
As can be seen in #1413 and similar previous fixes, we almost always
forget to call the addNoProxy function when adding a new bind.

At the moment, the ceremony around this is completely unnecessary
since we know that we are running all relevant components on the local
machine.  However, the current approach of each bind separately was
put in place when we thought that different services might be living
on different machines.

Adding sys.ip to NORPOXY in the scaffolding will mean that any service
using the scaffoldings can likely omit adding `addNoProxy` lines for
every service bind.

Signed-off-by: Steven Danna <steve@chef.io>