ci: use expeditor environment to get package idents (#1402)
In expeditor, we have access to the exact package identifiers
that were included in the build group. We can use these
identifiers to avoid eventual consistency issues we've seen with
the Habitat Depot.

We are currently unsure on the exact cause of the inconsistent
manifest creation. While we think it is the result of caching in
the Depot API, we have been unable to confirm this.

To facilitate hunting it down, I've also added logging for cases
where the expeditor environment doesn't match what is found in
the depot:

    2019-08-30 11:18:37 +0100 WARN: Found mismatch between sources:
    2019-08-30 11:18:37 +0100 WARN:    chef/automate-gateway/0.1.0/20190829202630 from expeditor environment
    2019-08-30 11:18:37 +0100 WARN:    chef/automate-gateway/0.1.0/20190829202631 from Habitat Depot
    2019-08-30 11:18:37 +0100 WARN: Using chef/automate-gateway/0.1.0/20190829202630

This script was growing rather large, so I've broken it up a bit:

- Added a number of PackageQuerier classes each of which knows
  how to look up packages from a single source. The
  MustExistQuerier is maybe a bridge too far, but oh well, I
  think the rest are fine.

- Added a ManifestGenerator class just to provide a line
  between "setup" logic and generation logic since there were a
  LOT of state variables in the main script.

- Added a logger so that we get timestamps in the logs

- Add some more log statements so we can see what configuration
  was used to generate a given manifest from looking at the logs.

Signed-off-by: Steven Danna <steve@chef.io>