adding an additional unit to capture from journalctl and removing a duplicate AddCommand line that was overwriting the previous entry (#1529)
For consistency with Chef Automate Frontends, the Automate Cluster uses chef-automate gather-logs in local fallback mode to capture logs for Support bundles on the backend nodes. This PR simply adds the hab-sup systemd unit to the command used to capture journalctl output. journalctl accepts multiple -u|--unit arguments without problem. With -u hab-sup we can also capture the systemd logs for the backend Habitat services running under the hab-sup unit.

The new modified command works seamlessly on front or backend nodes.

Additionally, this PR fixes a small bug, removing the following line that appears to be a duplicate:
g.AddCommand("journalctl_chef-automate", "journalctl", "--utc", "-u", "chef-automate")

I verified that in local fallback mode, the above line was overwriting the previous AddCommand entry and therefore never honoring the value of --log-lines even when the user specified it on the cli.

Signed-off-by: Jeremy J. Miller <jm@chef.io>