Applications sample data more useful for filtering use cases
Our sample data had not been updated since we added application and
environment and was using the service group name to encode this
information (which we recommend against doing) in several places.

The sample data now has an application with serveral services
(mobile-app-api) deployed to two environments which shows how searching
by environment and application can focus the UI on a single deployment.

There are at least two options for all filterable fields in the dataset.

Signed-off-by: Daniel DeLeo <dan@chef.io>