ci: temporary fix for bug in single-use buildkite queue
The single-use buildkite queue currently has a bug in which machines
are being reused. This poses a problem for some of our jobs since they
run and create files as root. Those files cannot later be cleaned up,
failing the build.

This PR adds a `sudo git clean` step to all such jobs in an attempt to
keep the rate of failures down. `sudo git clean` should clean out the
files that we would fail to clean out on a subsequent run.  We'll want
to pair this with a termination of all existing nodes in the single
use queue.

In the long run, we should consider reworking these tests so that they do
not require the single-use queue or sudo.

Signed-off-by: Steven Danna <steve@chef.io>