Bump Go version to 1.13 (#1546)
The Go version used by the build is controlled by the dependencies of
core/scaffolding-go, chef/scaffolding-go, and
chef/automate-scaffolding-go. Those packages have all been updated to
use Go 1.13. This GOLANG_VERSION file controls the version of Go that
gets installed in some build jobs and used in some local development
cases.

Merging this is going to force some developers to update their local
version of Go.

Bumping the Go version forces us to upgrade golangci-lint since we need
a version compiled with the same version of Go as the one we are using.
The new version of golangci-lint found a few more violations and added two linters.
I've disabled the new linters for now so those can be fixed in follow-up PRs.

Signed-off-by: Steven Danna <steve@chef.io>