Clean up IAM v2.1 version checking in UI (#1524)
* Clean up IAM v2.1 version checking in UI

1. Uses the new atLeastV2p1 selector in place of the individual
major/minor selectors, bringing the version determination to a
single point of truth, yielding simpler code.
2. Perhaps more importantly, abstracts the notion of "projects enabled"
to all consumers, so that they do not need to care about the specifics of the major and minor versions.
This allows for future version bumps to be much cleaner and simpler.

Now the code has clear distinction between projects
and things that are v1 vs. v2 (e.g. roles, aspects of teams, etc.).