Client Runs page, detecting bad page selection (#1553)
This change allows the client runs page to revert back to the first page of nodes if the page selected does not have nodes.

This problem was found with the IAM v2 global project filtering feature. When the project filter is updated it refreshes the page. Say the user is on page 2 with 100 nodes per page. If the user updates the project filter to have only 50 nodes then page 2 would be empty. Since the page 2 selection is in the URL after the refresh the user would be left with zero nodes.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>