automate-gateway: trigger detect job with newly-created ID
While investigating #1598 we observed that the creation of a new node
during the diagnostics tests was generating a detect job that always
failed to resolve any nodes.

The incoming node in the case of a Create does not have an ID because
the ID is created after calling the nodemanager service. It is this
newly created ID that should be used to trigger the detect job.  It
appears that BulkCreate already does the right thing. Update also uses
in.ID but in that case I believe the incoming node should already have
an ID set since it is an update.

Signed-off-by: Steven Danna <steve@chef.io>