Improve throughput with throttle instead of debounce
Spurred by @srenatus' question about debounce time,
I switched to the more suitable throttle time,
which emits at the start of the interval rather than the end.
This provided a bonus of removing a couple lines of code:
the throttling itself takes care of the staleness
for the IntrospectAll call.

Signed-off-by: michael sorens <msorens@chef.io>