Reduce even more network traffic
Realizing that there is typically a flurry of IntrospectSome calls
almost concurrent with the IntrospectAll call that fires once the
cache has expired, I realized that delaying staleness for the former
by just a couple seconds would eliminate those IntrospectSome calls.

Two components to this.
1. If an IntrospectAll call is fired just as the cache expires,
the nudge in the effects file would be sufficient.
2. But more typically, the IntrospectAll call would come some time later,
beyond the staleness-plus-nudge time.

In commit fe91d3d8, I removed the dependency
on the store's value of getLastFetchTime for IntrospectAll
specifically, which means that I am now free to move where that value
is set in the reducer to the *start* of the call rather than the
*completion* of the call. Net effect: as soon as IntrospectAll is
fired--which should occur before any IntrospectSome calls are fired--
that makes the cache appear fresh again to those IntrospectSome calls,
inhibiting them from making network calls.

Signed-off-by: michael sorens <msorens@chef.io>