lib/cereal: Add StartGuard to replace WaitGroup-based star checks
We were using WaitGroup to help ensure that we weren't calling Start()
wasn't called twice on various structs that manage long-lived go
routines.

This replaces those uses with a new type, StartGuard, that allows us
to customize the message that users (developers) will get when they
make this mistake.

We could have done this with recover, but this felt a big better.

Signed-off-by: Steven Danna <steve@chef.io>