Compliance reporting: Select date from the trend graph (#1575)
Clicking a column on the trend graph selects the focused date. I always have to look down at the trend graph to see where reports are and then go up to the date selector to choose that date. With this change, one does not need to use the date selector. Just click the date on the trend graph.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>