Use a cypress bind to make user mgmt test less flaky
Attempting to use a cy bind to make the test less flaky. Theory is other instances of 'Create User' on the page might be causing flakiness.

Also added a comment about paying attention to how you use data-cy in re-useable components. We should never be using the component in question in the same page twice so we should be fine here but a good thing to be aware of.

Signed-off-by: Tyler Cloke <tylercloke@gmail.com>