Fixing time zone problems (#1624)
Fixing a bug where the compliance reporting page would not allow the user in UTC+1 view the current day. The problem was the reportQuery state was being changed unintentionally. There was also some sections of code that when converting from moment to date was switching the current hour because of time zones.

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>