Fix type declarations in normalSort function and import into dropdown-component
Cleared out some of the typescript errors coming through that were due to
incorrect type declarations, mainly the input object should have been an
object and not an array.

I also imported the new normalSort function into the drowdown-component to
check that it was visually returning what I expect.  I'm still not passing
tests but I have a feeling that has to do with having a poorly written
test.