Refactor normalSort to accept input as an Array
Previously, the normalSort function was intaking an object, and then using Object.values() to
Pull out the values and convert to an Array.  After looking to use this across
the codebase, I realized that there are different ways a user may want to present their
input.  I've chosen to allow this up to the user to decide prior to using normalSort.