Projects in dropdown should be Normally sorted (#1620)
* WIP - Convert projectsArray to return naturally sorted list

Implemented the same use of sort combined with localeCompare
that is used in the policy-list.component so that the projectsArray
is easier to read in the dropdown. There are no tests yet associated
with this change.

* Add temp comments for reference

* Build mock object to test sorted outcome

* Develop proper way to mock @Input

* Add failing test cases

* Test cases passing and array is sorted normally

* Fix typescript warnings

An error was being thrown about the array of unsorted projets that I built
for testing in the spec file.  This was because an extra propery is assigned
on rendering, so I scaled it back to the same assignment it uses in the module
which is an {}.

* Fix typescript single quote warnings

* Remove excess code from spec file

Signed-off-by: seajamied <jdegnan@chef.io>

* Refactor tests to read simpler

Signed-off-by: seajamied <jdegnan@chef.io>

* Remove octothorpe from test describe block

* Revert prettier format spacing

* Revert one more prettier format

Signed-off-by: seajamied <jdegnan@chef.io>

* Fixed whitespace linter error

Signed-off-by: seajamied <jdegnan@chef.io>