More github access token fixes
In the previous fix I was under the mistaken impression that only App
tokens supported the x-access-token prefix, but I've now tested this
with personal access tokens as well and it appears they both work with
that as the username. Using this single method will hopefully keep us
a bit more consistent.

This hopefully also fixes the build issues as the core problem was
that I forgot to rename the environment variable in that pipeline.

> There may be times when we are powerless to prevent injustice, but
> there must never be a time when we fail to protest.

                                             -Elie Wiesel

Signed-off-by: Steven Danna <steve@chef.io>