Jamie/buttons hide profile description (#1666)
* Set max width on profile subheading

The subheading was running over into absolutely positioned buttons.  A max width of 610px
was set to contain the text.  610 was specifically chosen because it matches the width on
the compliance reporting page.

Signed-off-by: seajamied <jdegnan@chef.io>

* WIP - Unable to modify slot style

Need to learn how the slot content actually works so that I can understand how to style it properly

* Modify profile-details component layout

In order to be able to target the content injected by slot, I wrapped the slot in a div
with a class that gave me local access through the profile component SASS

Signed-off-by: seajamied <jdegnan@chef.io>

* Shrink max width to 520px

Signed-off-by: seajamied <jdegnan@chef.io>

* Alter max-width to use percentage

Signed-off-by: seajamied <jdegnan@chef.io>

* Revert formatting from prettier

Signed-off-by: seajamied <jdegnan@chef.io>

* Fix linting errors

Signed-off-by: seajamied <jdegnan@chef.io>