Use persistent state info for proj update state
This is a bug fix.
Previously, the project list page was using a local variable
based on the persistent state to determine the running/not-running
state that is used to dictate the content of the "Update Projects" button.
But that local variable was only being initialized when you
started the update. Thus, if you left the page, or if another user
visited the page, it would not be initialized correctly.

Signed-off-by: michael sorens <msorens@chef.io>