Suppress momentjs parse warning
```
WARN: 'Deprecation warning: value provided is not in a recognized RFC2822 or ISO format. moment construction falls back to js Date(), which is not reliable across all browsers and versions. Non RFC2822/ISO date formats are discouraged and will be removed in an upcoming major release. Please refer to http://momentjs.com/guides/#/warnings/js-date/ for more info.
```

The test is purposefully testing an invalid string passed to `moment`. The warning is expected.

Signed-off-by: Scott Christopherson <scott@chef.io>