datalifecycle: register correct workflow executor (#1752)
A recent CI failure revealed we were not correctly purging data. This
was the result of registering our workflow executor under the wrong
name. We were registering the workflow executor under the name of the
"workflow instance" rather than the "workflow".

This fix should result in the purge operation running as expected
again.

As follow up work we need to add a test for purge operations and make
it much harder to make this mistake.